+++ application/languages/en/mails.php
@@ -220,6 +220,14 @@ return [
     "paper_comment_by_editor_editor_copy_tpl_name" => "Editor's comment about a paper (editors copy)",
     "paper_comment_by_editor_editor_copy_mail_subject" => "%%REVIEW_CODE%% #%%ARTICLE_ID%% - editor comment about an article you're managing",
 
+    // Communication from author to assigned editors
+    "paper_comment_from_author_to_editor_editor_copy_tpl_name" => "Message from author to assigned editors (editors copy)",
+    "paper_comment_from_author_to_editor_editor_copy_mail_subject" => "%%REVIEW_CODE%% #%%ARTICLE_ID%% - new message about %%ARTICLE_RELATIONSHIP%%",
+
+    // Editor response to author message
+    "paper_editor_response_to_author_author_copy_tpl_name" => "Editor response to your message (author copy)",
+    "paper_editor_response_to_author_author_copy_mail_subject" => "%%REVIEW_CODE%% #%%ARTICLE_ID%% - an editor responded to your message",
+
     // Commentaire de l'auteur / lettre d'accompagnement
     'paper_author_comment_editor_copy_tpl_name' => "Author's comments / Cover letter",
     'paper_author_comment_editor_copy_mail_subject' => "Author's comments / Cover letter",